Database first connections
Le prgrm arrive à se connecter à la DB
On peut instancier et supprimer des produits de la DB (qui sont également des instances de classes dans le JAVA)
Il faut ajouter des contraintes au cas ou on essaie d'ajouter des produits existants et continuer sur les méthodes liées aux produis.
Aussi il faut upload la DB pour que @Louis-02-ESILV  puisse y accéder depuis son PC

+++ src/Accessoire.java
@@ -1,5 +1,5 @@
 public class Accessoire extends Produit{
-    public Accessoire(String nom, int prix, int nbex){
+    public Accessoire(String nom, double prix, int nbex){
         super(nom,prix,nbex);
     }
     @Override
