+++ src/Vetement.java
@@ -9,7 +9,7 @@ public class Vetement extends Produit{
         this.taille = taille;
     }
 
-    public Vetement(String nom, int prix, int nbex, int taille){
+    public Vetement(String nom, double prix, int nbex, int taille){
         super(nom,prix,nbex);//Gets the attributes from the extended class
         if(taille>=34&& taille <=54){this.taille=taille;}
         else{throw new IllegalArgumentException("Taille erronée");}
